requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
openai==2.8.0
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
Pillow==10.1.0
PyPDF2==3.0.1
tabula-py==2.9.0
openpyxl==3.1.2
lxml==4.9.3
pydantic==2.5.0
aiohttp==3.9.0
gunicorn==21.2.0
